requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
azure-cognitiveservices-speech==1.35.0
azure-core==1.33.0
azure-cosmos==4.7.0
azure-identity==1.18.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.6
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
marshmallow==3.26.1
msal==1.32.0
msal-extensions==1.3.1
multidict==6.3.2
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.40.0
orjson==3.10.16
packaging==24.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.2
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
yarl==1.18.3
uvicorn